pyproject.toml
[project]
name = "turboquant"
version = "0.1.0"
description = "LLM from scratch — modern architecture with quantization"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.5",
    "tokenizers>=0.21",
    "pyyaml>=6.0",
    "typer[all]>=0.15",
    "rich>=13.0",
    "safetensors>=0.5",
    "numpy>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.3",
    "ruff>=0.9",
    "mypy>=1.14",
]
wandb = ["wandb>=0.19"]

[project.scripts]
turboquant = "turboquant.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/turboquant"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
